Updated Pattern Recognition.

+++ src/map/MapCppTrackerRecon/MapCppTrackerRecon.cc
@@ -43,7 +43,7 @@ bool MapCppTrackerRecon::birth(std::string argJsonConfigDocument) {
 
   // Get the value above which an Exception is thrown
   assert(_configJSON.isMember("SciFiClustExcept"));
-  ClustException = _configJSON["SciFiClustExcept"].asDouble();
+  ClustException = _configJSON["SciFiClustExcept"].asInt();
 
   return true;
 }
